Major refactor (WIP)
