+++ common/src/main/java/dev/thomasglasser/tommylib/impl/mixin/minecraft/client/renderer/entity/player/PlayerRendererMixin.java
@@ -1,7 +1,7 @@
 package dev.thomasglasser.tommylib.impl.mixin.minecraft.client.renderer.entity.player;
 
 import com.mojang.blaze3d.vertex.PoseStack;
-import dev.thomasglasser.tommylib.TommyLib;
+import dev.thomasglasser.tommylib.impl.TommyLib;
 import dev.thomasglasser.tommylib.impl.GeckoLibUtils;
 import net.minecraft.client.model.PlayerModel;
 import net.minecraft.client.player.AbstractClientPlayer;
