+++ neoforge/src/main/java/dev/thomasglasser/tommylib/impl/client/TommyLibNeoForgeClient.java
@@ -1,6 +1,6 @@
 package dev.thomasglasser.tommylib.impl.client;
 
-import dev.thomasglasser.tommylib.TommyLib;
+import dev.thomasglasser.tommylib.impl.TommyLib;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
