+++ neoforge/src/main/java/dev/thomasglasser/tommylib/impl/data/lang/TommyLibEnUsLanguageProvider.java
@@ -1,6 +1,6 @@
 package dev.thomasglasser.tommylib.impl.data.lang;
 
-import dev.thomasglasser.tommylib.TommyLib;
+import dev.thomasglasser.tommylib.impl.TommyLib;
 import dev.thomasglasser.tommylib.api.data.lang.ExtendedEnUsLanguageProvider;
 import dev.thomasglasser.tommylib.api.tags.ConventionalBlockTags;
 import dev.thomasglasser.tommylib.api.tags.ConventionalItemTags;
