+++ neoforge/src/main/java/dev/thomasglasser/tommylib/impl/data/tags/TommyLibItemTagsProvider.java
@@ -1,6 +1,6 @@
 package dev.thomasglasser.tommylib.impl.data.tags;
 
-import dev.thomasglasser.tommylib.TommyLib;
+import dev.thomasglasser.tommylib.impl.TommyLib;
 import dev.thomasglasser.tommylib.api.data.tags.ExtendedItemTagsProvider;
 import dev.thomasglasser.tommylib.api.tags.ConventionalItemTags;
 import java.util.concurrent.CompletableFuture;
